requirements.txt
albumentations
black
flake8
isort
loguru
mkdocs
matplotlib
monai==1.3.2
monai-weekly==1.4.dev2431
nibabel
numpy==1.24.2
opencv-python
pip
plotly
python-dotenv
scikit-learn==1.5.1
scipy==1.13.1
seaborn
tqdm
typer
-e .
